requirements.txt
chumpy
neural_renderer_pytorch
tqdm
numpy
opendr
opencv_python
h5py
scipy
ipdb
tensorboardX
